refactor: Company 생성 반환 타입 변경
