+++ company-service/src/main/java/com/firstlogistics/companyservice/presentation/dto/response/CreateCompanyResponse.java
@@ -1,0 +1,13 @@
+package com.firstlogistics.companyservice.presentation.dto.response;
+
+import com.firstlogistics.companyservice.application.dto.result.CompanyResult;
+import java.util.UUID;
+
+public record CreateCompanyResponse(
+        UUID companyId
+) {
+
+    public static CreateCompanyResponse from(CompanyResult result) {
+        return new CreateCompanyResponse(result.id());
+    }
+}
